refactor(app): extract canvas layout callbacks + effects into useCanvasLayout (jn3)
Creates src/hooks/useCanvasLayout.ts (~366 lines) to own all story and route
canvas layout logic: applyStoryLayout, handleTidyUp, applyRouteLayout,
handleChangeStoryCanvasLayoutMode, handleChangeStoryCanvasGroupingMode,
handleChangeRouteCanvasLayoutMode, handleChangeRouteCanvasGroupingMode, and
the three layout-refresh/auto-center useEffects that fire after project load.

Hook receives explicit params: blocks/setBlocks, analysisResult,
routeAnalysisResult, routeNodeLayoutCache, pending layout refs,
projectSettings/updateProjectSettings, analysis pending flags, and
center-request dispatchers. Returns only the 7 handlers App.tsx needs.

Removes storyCanvasLayout + routeCanvasLayout computation imports from App.tsx
(now owned by the hook). App.tsx: 3000 → 2686 lines (−314). 686 tests pass.
