+++ .beads/interactions.jsonl
@@ -32,3 +32,5 @@
 {"id":"int-ab6afd3e","kind":"field_change","created_at":"2026-06-27T20:02:48.9159121Z","actor":"G.Hirpara","issue_id":"bmf-vangard-renpy-ide-i1s","extra":{"field":"status","new_value":"closed","old_value":"in_progress","reason":"handleOpenCharacterEditor/handleUpdateCharacter extracted into useCharacterManagement hook"}}
 {"id":"int-cf2974d5","kind":"field_change","created_at":"2026-06-27T20:02:49.5814437Z","actor":"G.Hirpara","issue_id":"bmf-vangard-renpy-ide-1es","extra":{"field":"status","new_value":"closed","old_value":"in_progress","reason":"Scene/imagemap handlers moved into useCompositionState; hook now accepts integration props (activeTabId, setOpenTabs, setActiveTabId, setHasUnsavedSettings)"}}
 {"id":"int-5ec55c4c","kind":"field_change","created_at":"2026-06-27T20:07:13.4185734Z","actor":"G.Hirpara","issue_id":"bmf-vangard-renpy-ide-3m9","extra":{"field":"status","new_value":"closed","old_value":"in_progress","reason":"handleCloseTab/processTabCloseRequest/handleCreate|Close|Move|Open*Split/DnD handlers extracted into useTabLifecycle hook; App.tsx at 3337 lines (down from 3864)"}}
+{"id":"int-fa019486","kind":"field_change","created_at":"2026-06-27T21:32:29.2825224Z","actor":"G.Hirpara","issue_id":"bmf-vangard-renpy-ide-3c6","extra":{"field":"status","new_value":"closed","old_value":"in_progress","reason":"DualPaneContext established. Phases 1-3 complete: dirty state, tab mgmt, lifecycle, openers all in context. Bug fix: hooks called from App.tsx now use explicit params not useDualPane(). App.tsx 3337→3000 lines."}}
+{"id":"int-47730ba4","kind":"field_change","created_at":"2026-06-27T21:32:30.1160118Z","actor":"G.Hirpara","issue_id":"bmf-vangard-renpy-ide-37q","extra":{"field":"status","new_value":"closed","old_value":"in_progress","reason":"StoryElementsPanel callbacks extracted to useStoryElementsPanel; image/audio metadata handlers folded into useAssetManagement. App.tsx 3259→3000 lines."}}
